WIP: Merge 'process' into 'modify'
