+++ clarite/__init__.py
@@ -1,7 +1,7 @@
 # flake8: noqa
 from ._version import get_versions
 
-from .modules import process, modify, plot, describe, analyze, load, survey
+from .modules import modify, plot, describe, analyze, load, survey
 
 __version__ = get_versions()['version']
 del get_versions
