+++ clarite/cli/__init__.py
@@ -1,3 +1,3 @@
 # flake8: noqa
 from .cli import entry_point
-from .cli import analyze_cli, describe_cli, load_cli, modify_cli, plot_cli, process_cli
+from .cli import analyze_cli, describe_cli, load_cli, modify_cli, plot_cli
