+++ clarite/cli/commands/__init__.py
@@ -4,4 +4,3 @@ from .describe import describe_cli
 from .load import load_cli
 from .modify import modify_cli
 from .plot import plot_cli
-from .process import process_cli
