+++ clarite/cli/commands/modify.py
@@ -1,8 +1,10 @@
 import click
+import pandas as pd
+
 from ...modules import modify
 from ...internal.utilities import _validate_skip_only
-from ..parameters import arg_data, option_output, option_skip, option_only
-from ..custom_types import save_clarite_data
+from ..parameters import arg_data, arg_output, option_output, option_skip, option_only, CLARITE_DATA, OUTPUT_FILE
+from ..custom_types import ClariteData, save_clarite_data
 
 
 @click.group(name='modify')
@@ -217,3 +219,79 @@ def transform_variable(data, output, variable, transform, new_name):
         data.df = data.df.drop(variable, axis='columns')
     # Save
     save_clarite_data(data, output)
+
+
+@modify_cli.command(help="Categorize data based on the number of unique values")
+@arg_data
+@option_output
+@click.option('--cat_min', default=3, help="Minimum number of unique values in a variable to make it a categorical type")
+@click.option('--cat_max', default=6, help="Maximum number of unique values in a variable to make it a categorical type")
+@click.option('--cont_min', default=15, help="Minimum number of unique values in a variable to make it a continuous type")
+def categorize(data, output, cat_min, cat_max, cont_min):
+    # Categorize and convert to ClariteData types
+    df_bin, df_cat, df_cont, df_check = modify.categorize(data.df, cat_min=cat_min, cat_max=cat_max, cont_min=cont_min)
+    data_bin = ClariteData(data.name + "_bin", df=df_bin)
+    data_cat = ClariteData(data.name + "_cat", df=df_cat)
+    data_cont = ClariteData(data.name + "_cont", df=df_cont)
+    data_check = ClariteData(data.name + "_check", df=df_check)
+    # Save Data
+    save_clarite_data(data=data_bin, output=f"{output}_bin")
+    save_clarite_data(data=data_cat, output=f"{output}_cat")
+    save_clarite_data(data=data_cont, output=f"{output}_cont")
+    save_clarite_data(data=data_check, output=f"{output}_check")
+
+
+@modify_cli.command(help="Merge variables from two different datasets into one")
+@click.argument('left', type=CLARITE_DATA)
+@click.argument('right', type=CLARITE_DATA)
+@arg_output
+@click.option('--how', '-h', default='outer', type=click.Choice(['left', 'right', 'inner', 'outer']), help="Type of Merge")
+def merge_variables(left, right, output, how):
+    # Merge
+    result = modify.merge_variables(left.df, right.df, how)
+    result = ClariteData(name=f"{left.name}.{right.name}", df=result)
+    # Save
+    save_clarite_data(result, output)
+
+
+@modify_cli.command(help="Merge rows from two different datasets into one")
+@click.argument('top', type=CLARITE_DATA)
+@click.argument('bottom', type=CLARITE_DATA)
+@arg_output
+def merge_rows(top, bottom, output):
+    # Merge
+    extra = set(list(top.df)) - set(list(bottom.df))
+    missing = set(list(bottom.df)) - set(list(top.df))
+    if len(extra) > 0:
+        raise ValueError(f"Couldn't merge rows: Extra columns in the 'bottom' data: {', '.join(extra)}")
+    elif len(missing) > 0:
+        raise ValueError(f"Couldn't merge rows: Missing columns in the 'bottom' data: {', '.join(missing)}")
+    elif (top.df.dtypes != bottom.df.dtypes).any():
+        raise ValueError("Couldn't merge rows: different data types")
+    else:
+        result = pd.concat([top.df, bottom.df], verify_integrity=True, sort=False)
+        result = ClariteData(name=f"{top.name}.{bottom.name}", df=result)
+    # Save
+    save_clarite_data(result, output)
+
+
+@modify_cli.command(help="Move variables from one dataset to another")
+@click.argument('left', type=CLARITE_DATA)
+@click.argument('right', type=CLARITE_DATA)
+@click.option('--output_left', default=None, type=OUTPUT_FILE)
+@click.option('--output_right', default=None, type=OUTPUT_FILE)
+@option_skip
+@option_only
+def move_variables(left, right, output_left, output_right, skip, only):
+    # Move
+    left_df, right_df = modify.move_variables(left.df, right.df, skip=skip, only=only)
+    # Save
+    # Modify in-place if output is not specified
+    if output_left is None:
+        output_left = left.name
+    if output_right is None:
+        output_right = right.name
+    left = ClariteData(name=left.name, df=left_df)
+    right = ClariteData(name=right.name, df=right_df)
+    save_clarite_data(left, output_left)
+    save_clarite_data(right, output_right)
